+++ qoqo_qiskit_devices/src/devices/ibm_lagos.rs
@@ -35,6 +35,9 @@ impl IBMLagosDeviceWrapper {
     /// Create a new IBMLagosDevice instance.
     #[new]
     pub fn new() -> Self {
+        Python::with_gil(|py| {
+            py.run("import warnings; warnings.warn(\"Device ibm_lagos has been retired. Setting noise information is not possible.\", category=DeprecationWarning, stacklevel=2)", None, None).unwrap();
+        });
         Self {
             internal: IBMLagosDevice::new(),
         }