+++ roqoqo_qiskit_devices/tests/integration/main.rs
@@ -13,4 +13,3 @@
 #[allow(deprecated)]
 #[cfg(test)]
 mod device;
-pub use device::*;