+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This changelog tracks changes of the qoqo_qiskit project starting at version 0.1.0 (initial release).
 
+### 0.2.3
+
+* Updated to qoqo 1.9
+* Added support for Python 3.12
+* Pinned qiskit dependency to <0.46
+* Added deprecation warnings for retired devices (Lagos, Nairobi, Perth)
+
 ### 0.2.2
 
 * Updated qoqo_qiskit_devices `get_decoherence_on_gate_model` method docstring